+++ app/src/main/java/com/example/education/presentation/fragments/homework1_1/SecondHoWo3Fragment.kt
@@ -1,9 +1,9 @@
-package com.example.education.presentation.fragments.homework3
+package com.example.education.presentation.fragments.homework1_1
 
 import androidx.fragment.app.Fragment
 import com.example.education.R
 
-class SecondHoWo3Fragment : Fragment(R.layout.fragment_second_ho_wo_3) {
+class SecondHoWo3Fragment : Fragment(R.layout.fragment_second_ho_wo_1_1) {
 
     companion object {
         const val SECOND_HO_WO_3_FRAGMENT_TAG = "SECOND_HO_WO_3_FRAGMENT_TAG"
