+++ app/src/main/java/com/example/education/presentation/fragments/homework1_1/ThirdHoWo3Fragment.kt
@@ -1,18 +1,18 @@
-package com.example.education.presentation.fragments.homework3
+package com.example.education.presentation.fragments.homework1_1
 
 import android.os.Bundle
 import android.view.View
 import androidx.fragment.app.Fragment
 import by.kirich1409.viewbindingdelegate.viewBinding
 import com.example.education.R
-import com.example.education.databinding.FragmentThirdHoWo3Binding
+import com.example.education.databinding.FragmentThirdHoWo11Binding
 import com.example.education.presentation.fragments.homework1.secondPage.SecondPageFragment
 
-class ThirdHoWo3Fragment : Fragment(R.layout.fragment_third_ho_wo_3) {
+class ThirdHoWo3Fragment : Fragment(R.layout.fragment_third_ho_wo_1_1) {
 
     private var countColor: Int = 1
 
-    private val viewBinding: FragmentThirdHoWo3Binding by viewBinding(FragmentThirdHoWo3Binding::bind)
+    private val viewBinding: FragmentThirdHoWo11Binding by viewBinding(FragmentThirdHoWo11Binding::bind)
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
