+++ app/src/main/java/com/example/education/presentation/fragments/homework3/ListPlanetsFragment.kt
@@ -1,4 +1,4 @@
-package com.example.education.presentation.fragments.homework4
+package com.example.education.presentation.fragments.homework3
 
 import android.os.Bundle
 import android.view.View
