+++ app/src/main/java/com/example/education/presentation/fragments/homework4/MultiItemTypeFragment.kt
@@ -1,4 +1,4 @@
-package com.example.education.presentation.fragments.homework5
+package com.example.education.presentation.fragments.homework4
 
 import android.os.Bundle
 import android.view.View
