+++ app/src/main/java/com/example/education/presentation/fragments/homework5/A2Fragment.kt
@@ -1,4 +1,4 @@
-package com.example.education.presentation.fragments.homework6
+package com.example.education.presentation.fragments.homework5
 
 import android.os.Bundle
 import android.view.View
