+++ app/src/main/res/layout/fragment_b.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".presentation.fragments.homework6.BFragment">
+    tools:context=".presentation.fragments.homework5.BFragment">
 
     <androidx.appcompat.widget.AppCompatEditText
         android:id="@+id/et_b"
