+++ app/src/main/res/layout/fragment_b_2.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".presentation.fragments.homework6.B2Fragment">
+    tools:context=".presentation.fragments.homework5.B2Fragment">
 
     <TextView
         android:id="@+id/tv_text_for_previous_page"
