+++ app/src/main/res/layout/fragment_c.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".presentation.fragments.homework6.CFragment">
+    tools:context=".presentation.fragments.homework5.CFragment">
 
     <androidx.appcompat.widget.AppCompatButton
         android:id="@+id/btn_to_c2"
